requirements.txt
streamlit>=1.31.0
opencv-python-headless>=4.9.0
numpy>=1.26.0
Pillow>=10.2.0




